code/requirements.txt
coremltools==7.1
matplotlib==3.8.1
numpy==1.26.2
pandas==2.1.4
PyYAML==5.1
PyYAML==6.0.1
scikit_learn==1.3.2
seaborn==0.13.0
tensorflow==2.14.0
tensorflow_macos==2.14.0